Refactor: split file
